+++ x/posts/keeper/keeper.go
@@ -8,20 +8,23 @@ import (
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+
 	"github.com/desmos-labs/desmos/x/posts/types"
+	"github.com/desmos-labs/desmos/x/relationships"
 )
 
 // Keeper maintains the link to data storage and exposes getter/setter methods for the various parts of the state machine
 type Keeper struct {
 	// The reference to the ParamsStore to get and set posts specific params
 	paramSubspace paramstypes.Subspace
+	relKeeper     relationships.Keeper // Relationships keeper to keep track of blocked users
 
 	StoreKey sdk.StoreKey       // Unexposed key to access store from sdk.Context
 	Cdc      *codec.LegacyAmino // The wire codec for binary encoding/decoding.
 }
 
 // NewKeeper creates new instances of the posts Keeper
-func NewKeeper(cdc *codec.LegacyAmino, storeKey sdk.StoreKey, paramSpace paramstypes.Subspace) Keeper {
+func NewKeeper(cdc *codec.LegacyAmino, storeKey sdk.StoreKey, paramSpace params.Subspace, relKeeper relationships.Keeper) Keeper {
 	if !paramSpace.HasKeyTable() {
 		paramSpace = paramSpace.WithKeyTable(types.ParamKeyTable())
 	}
@@ -30,12 +33,18 @@ func NewKeeper(cdc *codec.LegacyAmino, storeKey sdk.StoreKey, paramSpace paramst
 		StoreKey:      storeKey,
 		Cdc:           cdc,
 		paramSubspace: paramSpace,
+		relKeeper:     relKeeper,
 	}
 }
 
 // -------------
 // --- Posts
 // -------------
+
+// IsUserBlocked tells if the given blocker has blocked the given blocked user
+func (k Keeper) IsUserBlocked(ctx sdk.Context, blocker, blocked sdk.AccAddress) bool {
+	return k.relKeeper.IsUserBlocked(ctx, blocker, blocked)
+}
 
 // SavePost allows to save the given post inside the current context.
 // It assumes that the given post has already been validated.