+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Version 0.14.0
 ## Bug fixes
 - Fixed height not working in REST queries (#299)
+- Implemented blocked users checks (#298)
 
 # Version 0.13.0
 ## Changes